fix missing 'minimatch' module when build from CI, part 2
it appears that the problem still persists on remote
computers where the minimatch module is not installed globally.

in first part of the fix we made sure that npm installs
required module, and in this part we make sure that vsce
will actually pack this installed module into an extention.

this is done by changing what vsce should ignore when packing
an extension in .vscodeignore:
1. by default when packing vsce ignores all devDependencies,
   so we move `minipatch` to dependencies
2. we whitelist node_modules fully for vsce to be able to pack
   mentioned dependencies and not ignore them